Moved images, fonts and global styles to assets folder
