+++ src/pages/Login/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
 import { HeaderWelcome, Logo, LoginBox, LoginButton, Img, LoginPage, Content, Container } from './styles'
-import logo from '../../assets/logo.svg'
-import githubLogo from '../../assets/github.svg'
-import welcome from '../../assets/welcome.jpg'
+import logo from '../../assets/images/logo.svg'
+import githubLogo from '../../assets/images/github.svg'
+import welcome from '../../assets/images/welcome.jpg'
 
 import { AUTHORIZE_URL } from '../../constants'
 import { GITHUB_CLIENT_ID } from '../../env'
